pyproject.toml
[project]
name = "e-commers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.1",
    "django==5.2.4",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.1",
    "dotenv==0.9.9",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.10.1",
    "pyotp==2.9.0",
    "python-dotenv==1.1.1",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "uv==0.8.16",
]
